Fix Rendering System and add Origin rotation
You can now set an origin offset on the Transform Component. The
Rendering system will take this into account for rotations.

The Rendering system is no longer accidentally working either. Instead
it translates everything to Meters and applies the proper values when
rendering instead of awkwardly applying PPM to the scale and for
some reason mixing pixel and meter units.

(,o^-^)--0