Fixing TweenComponent bug on Reset
Bug on reset where Tweens were calling .free() in the reset() method,
which could cause a problem if this occurred mid cycle and the tween
manager didn't know about it. Instead we now pause and kill the tweens
and timeline so that we can let the TweenManager do its job and clean-up
our Tweens and release them back to the pool.